UI Card detail: code cleaning
