+++ ui/main/src/app/modules/calendar/calendar.component.html
@@ -1,4 +1,4 @@
-<!-- Copyright (c) 2018-2021, RTE (http://www.rte-france.com)              -->
+<!-- Copyright (c) 2018-2022, RTE (http://www.rte-france.com)              -->
 <!-- See AUTHORS.txt                                                       -->
 <!-- This Source Code Form is subject to the terms of the Mozilla Public   -->
 <!-- License, v. 2.0. If a copy of the MPL was not distributed with this   -->
@@ -20,7 +20,7 @@
     <ng-template #cardDetail let-modal>
       <div class="modal-body">
         <div>
-          <of-card [parentModalRef] = "modalRef" [screenSize]="'lg'" [displayContext]="displayContext"> </of-card>
+          <of-card [parentModalRef] = "modalRef" [screenSize]="'lg'" > </of-card>
         </div>
       </div>
   </ng-template>
