+++ ui/main/src/app/modules/card/card.component.html
@@ -7,8 +7,8 @@
 <!-- This file is part of the OperatorFabric project.                      -->
 
 <div *ngIf="card && cardState;" style="font-size:13px;" [hidden]="cardLoadingInProgress || cardNotFound">
-  <of-card-body [cardState]="cardState" [card]="card" [childCards]="childCards" [currentPath]="_currentPath"
-    [parentModalRef]="parentModalRef" [screenSize]="screenSize" [displayContext]="displayContext"
+  <of-card-body [cardState]="cardState" [card]="card" [childCards]="childCards"
+    [parentModalRef]="parentModalRef" [screenSize]="screenSize"
     [parentComponent]=this>
   </of-card-body>
 </div>
