+++ ui/main/src/app/modules/monitoring/components/monitoring-table/monitoring-table.component.html
@@ -42,7 +42,7 @@
 <ng-template #cardDetail let-modal>
     <div class="modal-body">
         <div>
-            <of-card [parentModalRef] = "modalRef" [screenSize]="'lg'" [displayContext]="displayContext"> </of-card>
+            <of-card [parentModalRef] = "modalRef" [screenSize]="'lg'"> </of-card>
         </div>
     </div>
 </ng-template>
